+++ src/Compiler/NodeVisitor/NodeVisitorStorage.php
@@ -4,31 +4,40 @@ declare(strict_types=1);
 
 namespace Efabrica\PHPStanLatte\Compiler\NodeVisitor;
 
+use Efabrica\PHPStanLatte\Compiler\NodeVisitor\Behavior\ExprTypeNodeVisitorInterface;
 use InvalidArgumentException;
 use PhpParser\NodeVisitor;
 
 final class NodeVisitorStorage
 {
-    /** @var array<int, NodeVisitor[]> */
-    private array $nodeVisitors = [];
+    private const WITHOUT_SCOPE = 'without_scope';
+
+    private const WITH_SCOPE = 'with_scope';
+
+    /** @var array<string, array<int, NodeVisitor[]>> */
+    private array $nodeVisitors = [
+        self::WITHOUT_SCOPE => [],
+        self::WITH_SCOPE => [],
+    ];
 
     public function addNodeVisitor(int $priority, NodeVisitor $nodeVisitor): void
     {
         if ($priority < 0 || $priority > 10000) {
             throw new InvalidArgumentException('Priority must be set between 0 and 10000');
         }
-        if (!isset($this->nodeVisitors[$priority])) {
-            $this->nodeVisitors[$priority] = [];
+        $scope = $nodeVisitor instanceof ExprTypeNodeVisitorInterface ? self::WITH_SCOPE : self::WITHOUT_SCOPE;
+        if (!isset($this->nodeVisitors[$scope][$priority])) {
+            $this->nodeVisitors[$scope][$priority] = [];
         }
-        $this->nodeVisitors[$priority][] = $nodeVisitor;
+        $this->nodeVisitors[$scope][$priority][] = $nodeVisitor;
     }
 
     /**
      * @return array<int, NodeVisitor[]>
      */
-    public function getNodeVisitors(): array
+    public function getNodeVisitors(bool $withScope = false): array
     {
-        $nodeVisitors = $this->nodeVisitors;
+        $nodeVisitors = $this->nodeVisitors[$withScope ? self::WITH_SCOPE : self::WITHOUT_SCOPE];
         ksort($nodeVisitors);
         return $nodeVisitors;
     }