+++ src/Rule/LatteTemplatesRule.php
@@ -157,7 +157,6 @@ final class LatteTemplatesRule implements Rule
 
             try {
                 $compileFilePath = $this->latteToPhpCompiler->compileFile($template, $context);
-                require($compileFilePath); // load type definitions from compiled template
             } catch (Throwable $e) {
                 $errors = array_merge($errors, $this->errorBuilder->buildErrors([new Error($e->getMessage() ?: get_class($e), $template->getPath())], $templatePath, null, $context));
                 continue;